User's groups fix
A/ If the GID exists, just add the user to that group and assume that's the
right thing to do.  It probably is.

B/ If the group exists (and has a different GID) then make a new group
called `creative_<group-name>` with the given GID and add the user to
that.

We used to only do B.  Now we do A, then B.